requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.5.2
bcrypt==5.0.0
cffi==1.17.1
click==8.1.8
cryptography==46.0.3
dnspython==2.6.1
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.0
fastapi==0.119.0
greenlet==3.1.1
h11==0.16.0
httptools==0.6.4
idna==3.11
importlib-metadata==8.5.0
importlib-resources==6.4.5
iniconfig==2.1.0
mako==1.3.10
MarkupSafe==2.1.5
packaging==25.0
passlib==1.7.4
pluggy==1.5.0
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.23
pydantic==2.10.6
pydantic-core==2.27.2
pydantic-settings==2.8.1
pytest==8.3.5
python-dotenv==1.0.1
python-jose==3.4.0
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.44
starlette==0.44.0
tomli==2.3.0
typing-extensions==4.13.2
uvicorn==0.33.0
uvloop==0.22.1
watchfiles==0.24.0
websockets==13.1
zipp==3.20.2
